requirements.txt
python-dotenv~=0.19.2
disnake
humanize~=4.0.0
disnake-jishaku
dblpy
motor~=2.5.1
topggpy
pymongo[srv]
git+https://github.com/DisnakeCommunity/disnake-ext-components
typing-extensions